refactor: rename core_logger to real_logger and instance-ize log_pattern
- Rename core_logger → real_logger (class, files, references)
- Refactor log_pattern from static utility to per-logger instance,
  eliminating global-state races during hot-reload
- Rename logger_proxy internal member real_logger → target_
- Fix socket_appender msg_len cast truncation
- Change json_object from std::map to std::unordered_map
- Fix docs/config key consistency in README and DESIGN.md
- Add malformed JSON config tests and log truncation tests
- Harden test macros with nullptr guards
- Clean up dead code and typos (.clang-tidy, code-format.py, build-rpm.sh)

+++ src/lib/appender/socket_appender.cpp
@@ -399,7 +399,7 @@ namespace log4cpp::appender {
             if (connection_fsm_state::ESTABLISHED != this->connection_state) {
                 return;
             }
-            ssize_t sent = send(this->sock_fd, msg, static_cast<int>(msg_len), 0);
+            ssize_t sent = send(this->sock_fd, msg, msg_len, 0);
             if (sent < 0) {
                 // Connection lost, notify reconnect thread
                 r_lock.unlock();
@@ -425,7 +425,7 @@ namespace log4cpp::appender {
         }
         else {
             // For UDP, just ignore the error
-            (void)send(this->sock_fd, msg, static_cast<int>(msg_len), 0);
+            (void)send(this->sock_fd, msg, msg_len, 0);
         }
     }
 } // namespace log4cpp::appender
